[Add] TMP Extension for UI data binding

+++ idle_RPG_Project/Assets/01_Scripts/Data/PlayerData.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using UnityEngine;
 
 [System.Serializable]
@@ -5,8 +6,8 @@ public class PlayerData
 {
     public string Name { get; set; }
     public int Level { get; set; }
-    public int Experience { get; set; }
-    public int Gold { get; set; }
+    public BigInteger Experience { get; set; }
+    public BigInteger Gold { get; set; }
     public int Gem { get; set; }
 
     public PlayerData()
